Changing some documentation, workflow for bad metrics types, and decoupling tests a bit more
